notebook/marimo/requirements.txt
duckdb
marimo
pandas<2.3
polars
sqlalchemy-cratedb
